Module Tezos_raw_protocol_005_PsBabyM1__Delegate_services.Endorsing_rights

type t = {

level : Tezos_raw_protocol_005_PsBabyM1.Alpha_context.Raw_level.t;

delegate : Tezos_protocol_environment_005_PsBabyM1__Environment.Signature.Public_key_hash.t;

slots : int list;

estimated_time : Tezos_raw_protocol_005_PsBabyM1.Alpha_context.Timestamp.t option;

}

val get : 'a Tezos_protocol_environment_005_PsBabyM1__Environment.RPC_context.#simple -> ?⁠levels:Tezos_raw_protocol_005_PsBabyM1.Alpha_context.Raw_level.t list -> ?⁠cycles:Tezos_raw_protocol_005_PsBabyM1.Alpha_context.Cycle.t list -> ?⁠delegates:Tezos_protocol_environment_005_PsBabyM1__Environment.Signature.public_key_hash list -> 'a -> t list Tezos_protocol_environment_005_PsBabyM1__Environment.Error_monad.shell_tzresult Tezos_protocol_environment_005_PsBabyM1__Environment.Lwt.t

Retrieves the delegates allowed to endorse a block.

By default, it gives the endorsement slots for bakers that have at least one in the next block.

Parameters levels and cycles can be used to specify the (valid) level(s) in the past or future at which the endorsement rights have to be returned. Parameter delegates can be used to restrict the results to the given delegates. Returns the list of endorsement slots. Also returns the minimal timestamps that correspond to these slots.

Timestamps are omitted for levels in the past, and are only estimates for levels later that the next block, based on the hypothesis that all predecessor blocks were baked at the first priority.