Module Tezos_raw_protocol_005_PsBabyM1__Roll_storage

Basic roll manipulation.

If storage related to roll (a.k.a. `Storage.Roll`) are not used outside of this module, this interface enforces the invariant that a roll is always either in the limbo list or in a contract list.